refactor: move ui-post-cart-homepage into /components/homepage
